requirements.txt
requests
python-dotenv
msal